Changes to make the name of the mod in factorio be the same as the name of the folder chosen to generate the script.

+++ cMain.h
@@ -188,7 +188,7 @@ private:
 	long long template_row_num;
 	long long template_row_count;
 
-	int row_to_move;
+	long long row_to_move;
 
 	bool mine_building_found;
